+++ R/lightgbm.R
@@ -101,7 +101,6 @@ train_lightgbm <- function(
     )
 
   args <- process_bagging(args)
-  args <- process_parallelism(args)
   args <- process_objective_function(args, x, y)
 
   args <- sort_args(args)
@@ -192,17 +191,6 @@ process_objective_function <- function(args, x, y) {
         args$objective <- "multiclass"
       }
     }
-  }
-
-  args
-}
-
-# supply the number of threads as num_threads in params, clear out
-# any other thread args that might be passed as main arguments
-process_parallelism <- function(args) {
-  if (!is.null(args["num_threads"])) {
-    args$num_threads <- args[names(args) == "num_threads"]
-    args[names(args) == "num_threads"] <- NULL
   }
 
   args