+++ NEWS.md
@@ -2,6 +2,8 @@
 
 * Increased the minimum R version to R 4.1.
 
+* Fixed bug where `num_threads` argument were ignored for lightgbm engine (#105).
+
 # bonsai 0.3.2
 
 * Resolves a test failure ahead of an upcoming parsnip release (#95).